+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## Unreleased
+
+**Release date:** `N/A`
+
+### Added
+
+- Shortcut factory to create FOxID's from List\<int\>.
+
 ## 1.2.1
 
 **Release date:** `2024-01-14`